refactor: rename to dag-index
