+++ src/lib/blockstore.js
@@ -20,7 +20,7 @@ const MAX_ENCODED_BLOCK_LENGTH = (1024 * 1024 * 2) + 39 + 61
 export class R2Blockstore {
   /**
    * @param {R2Bucket} dataBucket
-   * @param {import('./index/api').Index} index
+   * @param {import('./dag-index/api.js').Index} index
    */
   constructor (dataBucket, index) {
     this._dataBucket = dataBucket
