+++ src/middleware.js
@@ -4,8 +4,8 @@ import { CarReader } from '@ipld/car'
 import { parseCid, HttpError, toIterable } from '@web3-storage/gateway-lib/util'
 import { BatchingR2Blockstore } from './lib/blockstore.js'
 import { version } from '../package.json'
-import { ContentClaimsIndex } from './lib/index/content-claims.js'
-import { MultiCarIndex, StreamingCarIndex } from './lib/index/car.js'
+import { ContentClaimsIndex } from './lib/dag-index/content-claims.js'
+import { MultiCarIndex, StreamingCarIndex } from './lib/dag-index/car.js'
 import { CachingBucket, asSimpleBucket } from './lib/bucket.js'
 import { MAX_CAR_BYTES_IN_MEMORY, CAR_CODE } from './constants.js'
